Adds pinch gesture override to Allegro deployment
Implements an index/thumb pinch gesture override for the Allegro hand deployment.

The system subscribes to an `index_pinch_{left,right}` topic, and when a pinch is detected, it overrides the thumb and index finger joint commands to predefined pinch positions.

This introduces new nodes (`*_pinch.py`) for both single and dual hand deployment.
A new node also detects when the index and thumb are close.